move interfaces to one file because otherwise it doesn't build, move limit apis into a separate interface
